refactor: Extract OutputMatcher class in common
