+++ velox/common/testutil/CMakeLists.txt
@@ -16,5 +16,9 @@ velox_add_library(velox_test_util ScopedTestTime.cpp TestValue.cpp)
 velox_link_libraries(velox_test_util PUBLIC velox_exception)
 
 if(${VELOX_BUILD_TESTING})
+  velox_add_library(velox_test_output_matcher OutputMatcher.cpp)
+  velox_link_libraries(velox_test_output_matcher PUBLIC Folly::folly
+                                                        GTest::gtest re2::re2)
+
   add_subdirectory(tests)
 endif()
